changing the makefile flag to content Wall Werror Wextra and fixing the error of the compilation with thoses flags

+++ C08/ex00/src/main.cpp
@@ -2,7 +2,7 @@
 
 void DisplayVecContent(const std::vector<int>& VectorTest)
 {
-    for(int i = 0; i < VectorTest.size(); i++)
+    for(size_t i = 0; i < VectorTest.size(); i++)
     {
         std::cout << VectorTest[i];
         std::cout << " ";
